Some minor cleanup of header include and declarations
- Moved Fusion::manage from cloner.h to fusion.h
- Moved FusionGuard to its own file to allow only include FusionGuard
  without Fusion
- Removed include of all_nodes.h from builder.h
